requirements.txt
tensorflow==2.15.0
tensorflow-addons==0.23.0
keras-cv==0.6.4
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.4
imbalanced-learn==0.11.0
opencv-python==4.8.1.78
Pillow==10.1.0
pydicom==2.4.3
albumentations==1.3.1
tf-explain==0.3.1
grad-cam==1.4.8
shap==0.44.0
matplotlib==3.8.2
seaborn==0.13.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
tqdm==4.66.1
loguru==0.7.2
pytest==7.4.3
httpx==0.25.2
